Moved random flows to seperate file
